+++ src/main/java/com/app/estore/repository/VendorRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface VendorRepository extends JpaRepository<Vendor, Integer> {
+public interface VendorRepository extends JpaRepository<Vendor, Long> {
     Optional<Vendor> findByEmail(String email);
 }
